+++ example/lib/main.dart
@@ -17,6 +17,7 @@ class MyApp extends StatelessWidget {
     onPreviousMinDateTapped: (date) {},
     onAfterMaxDateTapped: (date) {},
     weekdayStart: DateTime.monday,
+    initialFocusDate: DateTime(2023, 5),
     // initialDateSelected: DateTime(2022, 3, 15),
     // endDateSelected: DateTime(2022, 3, 20),
   );